+++ lib/src/view/custom_currency_picker.dart
@@ -1,5 +1,6 @@
 import 'package:currency_widget/currency_widget.dart';
 import 'package:currency_widget/src/assets/currencies_names/currencies_names.dart';
+import 'package:currency_widget/src/utils/currency_format_utils.dart';
 import 'package:currency_widget/src/utils/currency_picker_utils.dart';
 import 'package:currency_widget/src/utils/masked_text_editing_controller.dart';
 import 'package:flutter/material.dart';
@@ -51,10 +52,11 @@ class _CustomCurrencyPickerState extends State<CustomCurrencyPicker> {
       widget.currencyController.currency = _currencies[0];
     }
 
+    final mount = widget.currencyController.mount.value;
     controller = TextEditingController(
-      text: widget.currencyController.mount.value != null &&
-              widget.currencyController.mount.value! > 0
-          ? widget.currencyController.mount.value.toString()
+      text: mount != null && mount > 0
+          ? CurrencyFormatUtils.formatAmount(
+              mount, widget.currencyController.currency)
           : '',
     );
   }
@@ -73,7 +75,8 @@ class _CustomCurrencyPickerState extends State<CustomCurrencyPicker> {
           widget.currencyController.currency = _currencies[0];
           final mount = widget.currencyController.mount.value;
           if (mount != null && mount > 0) {
-            controller.text = mount.toStringAsFixed(_currencies[0].decimalDigits);
+            controller.text =
+                CurrencyFormatUtils.formatAmount(mount, _currencies[0]);
           }
         }
       });
@@ -122,8 +125,11 @@ class _CustomCurrencyPickerState extends State<CustomCurrencyPicker> {
                   widget.currencyController.mount.value = 0;
                   return;
                 }
+                final currency = widget.currencyController.currency;
                 try {
-                  String value = str.replaceAll(',', '');
+                  String value = str
+                      .replaceAll(currency.thousandSeparator, '')
+                      .replaceAll(currency.decimalSeparator, '.');
                   widget.currencyController.mount.value = double.parse(value);
                 } catch (e) {
                   widget.currencyController.mount.value = 0;
@@ -131,8 +137,13 @@ class _CustomCurrencyPickerState extends State<CustomCurrencyPicker> {
               },
               inputFormatters: [
                 AutoDecimalNumberFormatter(
-                    decimalDigits:
-                        widget.currencyController.currency.decimalDigits),
+                  decimalDigits:
+                      widget.currencyController.currency.decimalDigits,
+                  thousandSeparator:
+                      widget.currencyController.currency.thousandSeparator,
+                  decimalSeparator:
+                      widget.currencyController.currency.decimalSeparator,
+                ),
               ],
             ),
           ),
@@ -158,7 +169,7 @@ class _CustomCurrencyPickerState extends State<CustomCurrencyPicker> {
       widget.currencyController.currency = selected;
       final mount = widget.currencyController.mount.value;
       if (mount != null && mount > 0) {
-        controller.text = mount.toStringAsFixed(selected.decimalDigits);
+        controller.text = CurrencyFormatUtils.formatAmount(mount, selected);
       }
     });
   }